Handle quotes in Contentful rich text
