+++ _includes/layouts/home.njk
@@ -60,7 +60,7 @@ layout: layouts/base.njk
           </a>
           {% if note.authorCommentary %}
           <div class="text-foreground text-sm mb-4">
-            {{ note.authorCommentary | renderRichTextAsHtml | safe }}
+            {{ note.authorCommentary | safe }}
           </div>
           {% endif %}
